Abstract identity-claims package tests into separate files
